reanmed and moved to Propel SQL count vs PHP Count
